LibC: Remove a bunch of weak `pthread_*` symbols
