+++ Userland/Libraries/LibC/bits/stdio_file_implementation.h
@@ -10,6 +10,7 @@
 #include <LibC/bits/FILE.h>
 #include <LibC/bits/pthread_integration.h>
 #include <LibC/bits/wchar.h>
+#include <pthread.h>
 #include <sys/types.h>
 
 #pragma once
@@ -21,7 +22,7 @@ public:
         , m_mode(mode)
     {
         pthread_mutexattr_t attr = { __PTHREAD_MUTEX_RECURSIVE };
-        __pthread_mutex_init(&m_mutex, &attr);
+        pthread_mutex_init(&m_mutex, &attr);
     }
     ~FILE();
